Renamed page component files
