+++ src/pages/RegisterPage.js
@@ -1,0 +1,16 @@
+import { useEffect } from "react";
+import RegisterForm from "../components/RegisterForm";
+
+function Register() {
+	useEffect(() => {
+		document.title = "WCC - Register";
+	}, []);
+
+	return (
+		<div className="container">
+			<RegisterForm />
+		</div>
+	);
+}
+
+export default Register;
